Move integration and end-to-end tests into separate packages
