+++ src/functionalTest/java/uk/gov/hmcts/reform/pbis/integration/ServiceBusClientTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.pbis;
+package uk.gov.hmcts.reform.pbis.integration;
 
 import static java.util.stream.Collectors.toList;
 import static org.assertj.core.api.Assertions.assertThat;
